Added saving to SQLite db

+++ 2sem/09.02.16-intentservice/app/src/main/java/ru/ilyamodder/intentservicetest/classes/Weather.java
@@ -16,13 +16,13 @@ public class Weather implements Serializable {
         @SerializedName("main")
         private WeatherData mData;
         @SerializedName("dt")
-        private int mDatestamp;
+        private long mDatestamp;
 
         public WeatherData getData() {
             return mData;
         }
 
-        public int getDatestamp() {
+        public long getDatestamp() {
             return mDatestamp;
         }
     }
